Moved util/proc.go to util/proc/proc.go, added boilerplate
